requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
filelock==3.7.0
huggingface-hub==0.6.0
idna==3.3
loguru==0.6.0
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pip==21.3.1
protobuf==3.20.1
pyparsing==3.0.9
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sentencepiece==0.1.96
setuptools==60.2.0
tokenizers==0.12.1
torch==1.11.0+cu115
torchaudio==0.11.0+cu115
torchvision==0.12.0+cu115
tqdm==4.64.0
transformers==4.19.2
typing-extensions==4.2.0
urllib3==1.26.9
wheel==0.37.1
pycuda==2021.1
cuda-python==11.6.1
nvidia-tensorrt==8.2.3.0
onnxruntime-gpu==1.11.1
